requirements.txt
pymupdf=1.26.4
pillow=11.3.0
streamlit=1.48.1
